refactor: Implement toggleEditing function in adminHome page
The code changes in the adminHome page implement the toggleEditing function, which was previously missing. This function is responsible for toggling the editing mode in the adminHome page. It throws an error for now, indicating that it needs to be implemented. This commit adds the necessary function to the page.