feat: Fetch gallery images from API in Home component
This commit modifies the Home component to fetch gallery images from the API using an axios GET request. The fetched images are then set in the galleryImages state using the setGalleryImages function. This enhancement improves the functionality of the application by dynamically populating the gallery with images from the API.