feat: Add lazy loading for images in product detail page
This commit adds lazy loading functionality for images in the product detail page. It imports the `Link` component from `next/link` and the `FaShoppingCart` icon from `react-icons/fa`. When a user selects a size and clicks the "Add to Cart" button, the product is added to the cart and a success toast notification is displayed. If no size is selected, an error toast notification is shown. These enhancements improve the user experience and performance of the product detail page.