feat: Add lazy loading for images in multiple pages
This commit adds lazy loading functionality for images in the orders page, editProduct page, and gallery page. It improves the performance and security of the affected pages by importing the `useRouter` hook from `next/navigation` to handle navigation and checking for a valid token in the local storage. Additionally, it enhances the user experience by redirecting users to the signin page if no token is found. These changes optimize the rendering of images and provide a smoother browsing experience.